pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Nika <vlasova_nika69@mail.ru>"]
readme = "README.md"
packages = [
    {include = "gendiff"}
]

[tool.poetry.dependencies]
python = "^3.10.6"
PyYAML = "^6.0.1"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
